+++ android/src/main/AndroidManifest.xml
@@ -14,6 +14,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Android">
         <activity
+            android:name=".intents.IntentsActivity"
+            android:exported="false" />
+        <activity
             android:name=".mvp.MvpActivity"
             android:exported="false" />
         <activity
